Reduce all the numbers to smallest size

+++ Player.cpp
@@ -2,12 +2,12 @@
 #include "Player.h"
 
 
-Player::Player(int pin) {
+Player::Player(byte pin) {
   pinMode(pin, OUTPUT);
   _pin = pin;
 };
 
-void Player::play(int pitch, int duration) {
+void Player::play(short pitch, short duration) {
   tone(_pin, pitch, duration);
    // to distinguish the notes, set a minimum time between them.
   // the note's duration + 30% seems to work well:
@@ -18,16 +18,14 @@ void Player::play(int pitch, int duration) {
 };
 
 
-void Player::playTune(int tuneLength, int tune[][2]) {
+void Player::playTune(byte tuneLength, byte tune[][2]) {
  //output an frequency
- for(int thisNote=0;thisNote<tuneLength;thisNote++)
+ for(byte thisNote=0;thisNote<tuneLength;thisNote++)
  {
   // to calculate the note duration, take one second divided by the note type.
   //e.g. quarter note = 1000 / 4, eighth note = 1000/8, etc.
-  int pitch = tune[thisNote][0];
-  int duration = 1000 / tune[thisNote][1];
+  short pitch = tune[thisNote][0];
+  short duration = 1000 / tune[thisNote][1];
   play(pitch, duration);
  }
 };
-
-
